[#54] chore: SwiftData.Dependency 수정
구현체 구현에 따른 수정
